Add 256 to int and uint

+++ src/DssExecLib.sol
@@ -56,7 +56,7 @@ interface DssVat {
     function nope(address) external;
     function ilks(bytes32) external returns (uint256 Art, uint256 rate, uint256 spot, uint256 line, uint256 dust);
     function Line() external view returns (uint256);
-    function suck(address, address, uint) external;
+    function suck(address, address, uint256) external;
 }
 
 interface ClipLike {
@@ -76,8 +76,8 @@ interface JoinLike {
     function ilk() external returns (bytes32);
     function gem() external returns (address);
     function dec() external returns (uint256);
-    function join(address, uint) external;
-    function exit(address, uint) external;
+    function join(address, uint256) external;
+    function exit(address, uint256) external;
 }
 
 // Includes Median and OSM functions
